beauty

+++ prepare.sh
@@ -11,7 +11,7 @@ KVM_TEMPLATE_NAME="template.ubuntu_1604"
 ISO_SAVE="$KVM_HOME/boot/$ISO_NAME.iso"
 MNT_TMP=$(mktemp -d)
 
-if [[ ! -f $ISO_HTTP_NAME ]]; then
+if [[ ! -f $ISO_HTTP_NAME && ! -f $ISO_SAVE ]]; then
     curl -O $ISO_HTTP
 fi
 
@@ -22,8 +22,8 @@ if [[ ! -f $ISO_SAVE ]]; then
     umount $MNT_TMP
     rm -rf $MNT_TMP
     
-    apt-get update
-    apt-get -y install qemu-kvm libvirt-bin bridge-utils virt-manager ansible mkisofs libguestfs-tools
+#    apt-get update
+#    apt-get -y install qemu-kvm libvirt-bin bridge-utils virt-manager ansible mkisofs libguestfs-tools
     
     cp -r unattend_ubnt/* $ISO_TMP/
     
@@ -36,7 +36,7 @@ fi
 if [[ ! -f "$KVM_HOME/images/$KVM_TEMPLATE_NAME.img" ]]; then    
     virt-install --virt-type=kvm --hvm --os-variant=ubuntu16.04 --name $KVM_TEMPLATE_NAME --ram 2048 --vcpus=2 --graphics vnc,listen=0.0.0.0,password=Inst@ll --cdrom=$ISO_SAVE --network network=default,model=virtio --disk path=$KVM_HOME/images/$KVM_TEMPLATE_NAME.img,size=16,bus=virtio
 
-    sleep 10
+    sleep 30
     virsh shutdown $KVM_TEMPLATE_NAME
     virsh autostart --disable $KVM_TEMPLATE_NAME
     sleep 5
@@ -44,7 +44,7 @@ if [[ ! -f "$KVM_HOME/images/$KVM_TEMPLATE_NAME.img" ]]; then
     virsh undefine $KVM_TEMPLATE_NAME
     #Backup
     #cp $KVM_HOME/images/$KVM_TEMPLATE_NAME.img $KVM_HOME/images/$KVM_TEMPLATE_NAME.img.back
-    virt-sysprep --enable=cron-spool,dhcp-client-state,dhcp-server-state,logfiles,mail-spool,net-hwaddr,rhn-systemid,ssh-hostkeys,udev-persistent-net,utmp,yum-uuid -a $KVM_HOME/images/$KVM_TEMPLATE_NAME.img
+    virt-sysprep --enable cron-spool,dhcp-client-state,dhcp-server-state,logfiles,mail-spool,net-hwaddr,rhn-systemid,ssh-hostkeys,udev-persistent-net,utmp,yum-uuid -a $KVM_HOME/images/$KVM_TEMPLATE_NAME.img
 
 fi
